Finalising refactor.

+++ src/World/animationoverlay.cs
@@ -13,9 +13,9 @@ namespace Underworld
         public AnimationOverlay(int _index)
         {
             index = _index;
-            switch ((byte)GameConfig.GameSelected)
+            switch (GameConfig.GameSelected)
             {
-                case (byte)Game.Uw2:
+                case Game.Uw2:
                     PTR = 0x7c06 + 2 + (_index * 6); break; // Located after end of tilemap data;
                 default:
                     PTR = _index * 6; break;
@@ -29,9 +29,9 @@ namespace Underworld
         {
             get
             {
-                switch ((byte)GameConfig.GameSelected)
+                switch (GameConfig.GameSelected)
                 {
-                    case (byte)Game.Uw2: // data is at the end of the tilemap
+                    case Game.Uw2: // data is at the end of the tilemap
                         return (int)((Loader.getAt(UWTileMap.current_tilemap.lev_ark_block.Data, PTR, 16) >> 6) & 0x3ff);
                     default://but UW1 stores the data in it's own block
                         return (int)((Loader.getAt(UWTileMap.current_tilemap.ovl_ark_block.Data, PTR, 16) >> 6) & 0x3ff);
@@ -40,9 +40,9 @@ namespace Underworld
             set
             {
                 var newvalue = (value & 0x3FF) << 6;
-                switch ((byte)GameConfig.GameSelected)
+                switch (GameConfig.GameSelected)
                 {
-                    case (byte)Game.Uw2: // data is at the end of the tilemap
+                    case Game.Uw2: // data is at the end of the tilemap
                         {
                             var currentVal = (int)Loader.getAt(UWTileMap.current_tilemap.lev_ark_block.Data, PTR, 16);
                             currentVal &= 0x3F; //mask out the link to 0.
@@ -89,9 +89,9 @@ namespace Underworld
         {
             get
             {
-                switch ((byte)GameConfig.GameSelected)
+                switch (GameConfig.GameSelected)
                 {
-                    case (byte)Game.Uw2: // data is at the end of the tilemap
+                    case Game.Uw2: // data is at the end of the tilemap
                         return (int)Loader.getAt(UWTileMap.current_tilemap.lev_ark_block.Data, PTR + 2, 16);
                     default://but UW1 stores the data in it's own block
                         return (int)Loader.getAt(UWTileMap.current_tilemap.ovl_ark_block.Data, PTR + 2, 16);
@@ -99,9 +99,9 @@ namespace Underworld
             }
             set
             {
-                switch ((byte)GameConfig.GameSelected)
+                switch (GameConfig.GameSelected)
                 {
-                    case (byte)Game.Uw2: // data is at the end of the tilemap
+                    case Game.Uw2: // data is at the end of the tilemap
                         Loader.setAt(UWTileMap.current_tilemap.lev_ark_block.Data, PTR + 2, 16, value); break;
                     default://but UW1 stores the data in it's own block
                         Loader.setAt(UWTileMap.current_tilemap.ovl_ark_block.Data, PTR + 2, 16, value); break;
@@ -116,9 +116,9 @@ namespace Underworld
         {
             get
             {
-                switch ((byte)GameConfig.GameSelected)
+                switch (GameConfig.GameSelected)
                 {
-                    case (byte)Game.Uw2: // data is at the end of the tilemap
+                    case Game.Uw2: // data is at the end of the tilemap
                         return UWTileMap.current_tilemap.lev_ark_block.Data[PTR + 4] & 0x3f;
                     default://but UW1 stores the data in it's own block
                         return UWTileMap.current_tilemap.ovl_ark_block.Data[PTR + 4] & 0x3f;
@@ -126,9 +126,9 @@ namespace Underworld
             }
             set
             {
-                switch ((byte)GameConfig.GameSelected)
+                switch (GameConfig.GameSelected)
                 {
-                    case (byte)Game.Uw2: // data is at the end of the tilemap
+                    case Game.Uw2: // data is at the end of the tilemap
                         UWTileMap.current_tilemap.lev_ark_block.Data[PTR + 4] = (byte)(value & 0x3f);
                         break;
                     default://but UW1 stores the data in it's own block
@@ -145,9 +145,9 @@ namespace Underworld
         {
             get
             {
-                switch ((byte)GameConfig.GameSelected)
+                switch (GameConfig.GameSelected)
                 {
-                    case (byte)Game.Uw2: // data is at the end of the tilemap
+                    case Game.Uw2: // data is at the end of the tilemap
                         return UWTileMap.current_tilemap.lev_ark_block.Data[PTR + 5] & 0x3f;
                     default://but UW1 stores the data in it's own block
                         return UWTileMap.current_tilemap.ovl_ark_block.Data[PTR + 5] & 0x3f;
@@ -155,9 +155,9 @@ namespace Underworld
             }
             set
             {
-                switch ((byte)GameConfig.GameSelected)
+                switch (GameConfig.GameSelected)
                 {
-                    case (byte)Game.Uw2: // data is at the end of the tilemap
+                    case Game.Uw2: // data is at the end of the tilemap
                         UWTileMap.current_tilemap.lev_ark_block.Data[PTR + 5] = (byte)(value & 0x3f);
                         break;
                     default://but UW1 stores the data in it's own block
