+++ src/interaction/damage.cs
@@ -237,16 +237,16 @@ namespace Underworld
                         if ((objToDestroy.OneF0Class == 0) && (damagetype != 8))
                         {//weapons
                             if (
-                                (objToDestroy.item_id == 3) && ((byte)GameConfig.GameSelected == (byte)Game.Uw2)
+                                (objToDestroy.item_id == 3) && (GameConfig.GameSelected == Game.Uw2)
                                 ||
-                                (objToDestroy.item_id == 0x10) && ((byte)GameConfig.GameSelected == (byte)Game.Uw2)
+                                (objToDestroy.item_id == 0x10) && (GameConfig.GameSelected == Game.Uw2)
                                 )
                             {
                                 Debris = 0xC7;
                             }
                             else
                             {
-                                if ((byte)GameConfig.GameSelected == (byte)Game.Uw2)
+                                if (GameConfig.GameSelected == Game.Uw2)
                                 {
                                     Debris = 0xC5 + weaponObjectDat.skill(objToDestroy.item_id);
                                 }
@@ -269,7 +269,7 @@ namespace Underworld
                             }
                             else
                             {
-                                if (((byte)GameConfig.GameSelected == (byte)Game.Uw2) && (objToDestroy.item_id == 0x116))
+                                if ((GameConfig.GameSelected == Game.Uw2) && (objToDestroy.item_id == 0x116))
                                 {
                                     djinnbottle.DestroyDjinnBottle(objToDestroy, WorldObject);
                                 }
@@ -311,7 +311,7 @@ namespace Underworld
 
             if (Debris <= -1)
             {
-                if ((byte)GameConfig.GameSelected == (byte)Game.Uw2)
+                if (GameConfig.GameSelected == Game.Uw2)
                 {
                     Debris = GetObjectTypeDebris(objToDestroy, damagetype);
                 }
@@ -360,7 +360,7 @@ namespace Underworld
             {
                 if (objToDestroy.OneF0Class == 0x15)
                 {//wood chips
-                    if ((byte)GameConfig.GameSelected == (byte)Game.Uw2)
+                    if (GameConfig.GameSelected == Game.Uw2)
                     {
                         return 0xDC;
                     }
@@ -377,13 +377,13 @@ namespace Underworld
             }
             else
             {
-                if ((objToDestroy.item_id == 3) && ((byte)GameConfig.GameSelected == (byte)Game.Uw2))
+                if ((objToDestroy.item_id == 3) && (GameConfig.GameSelected == Game.Uw2))
                 {
                     return 0xC7;//broken dagger in UW2
                 }
                 else
                 {
-                    if ((byte)GameConfig.GameSelected == (byte)Game.Uw2)
+                    if (GameConfig.GameSelected == Game.Uw2)
                     {
                         return 0xC5 + weaponObjectDat.skill(objToDestroy.item_id);
                     }
@@ -398,7 +398,7 @@ namespace Underworld
 
         public static int ScaleDamage(int item_id, ref int basedamage, int damagetype)
         {
-            if ((byte)GameConfig.GameSelected == (byte)Game.Uw2)
+            if (GameConfig.GameSelected == Game.Uw2)
             {
                 return ScaleDamageUW2(item_id: item_id, basedamage: ref basedamage, damagetype: damagetype);
             }
