+++ src/interaction/trapdisarming.cs
@@ -24,7 +24,7 @@ namespace Underworld
 
                     if (foundtrigger != null)
                     {
-                        if ((byte)GameConfig.GameSelected == (byte)Game.Uw2)
+                        if (GameConfig.GameSelected == Game.Uw2)
                         {
                             if (foundtrigger.minorclass == 3)
                             {//found a (movement/action) based trigger
@@ -77,7 +77,7 @@ namespace Underworld
         /// <returns></returns>
         static playerdat.SkillCheckResult DoTrapSkillCheck(int searchskill)
         {
-            if ((byte)GameConfig.GameSelected == (byte)Game.Uw2)
+            if (GameConfig.GameSelected == Game.Uw2)
             {
                 var difficulty = 0xA + (worlds.GetWorldNo(playerdat.dungeon_level) << 1);
                 return playerdat.SkillCheck(searchskill, difficulty, true);
@@ -107,7 +107,7 @@ namespace Underworld
 
                     if (foundtrigger != null)
                     {
-                        if ((byte)GameConfig.GameSelected == (byte)Game.Uw2)
+                        if (GameConfig.GameSelected == Game.Uw2)
                         {
                             if (foundtrigger.minorclass == 3)
                             {//found a (movement/action) based trigger
