+++ src/objectdata/triggerobjectdat.cs
@@ -30,7 +30,7 @@ namespace Underworld
         {
             get
             {
-                if ((byte)GameConfig.GameSelected==(byte)Game.Uw2)
+                if (GameConfig.GameSelected == Game.Uw2)
                 {
                     return triggertypes.UNLOCK_UW2;
                 }
@@ -44,7 +44,7 @@ namespace Underworld
         {
             get
             {
-                if ((byte)GameConfig.GameSelected==(byte)Game.Uw2)
+                if (GameConfig.GameSelected == Game.Uw2)
                 {
                     return triggertypes.OPEN_UW2;
                 }
